src/main/webapp/app/entities/bon-reception-item/route/bon-reception-item-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IBonReceptionItem } from '../bon-reception-item.model';
import { BonReceptionItemService } from '../service/bon-reception-item.service';

import { BonReceptionItemRoutingResolveService } from './bon-reception-item-routing-resolve.service';

describe('BonReceptionItem routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: BonReceptionItemRoutingResolveService;
  let service: BonReceptionItemService;
  let resultBonReceptionItem: IBonReceptionItem | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(BonReceptionItemRoutingResolveService);
    service = TestBed.inject(BonReceptionItemService);
    resultBonReceptionItem = undefined;
  });

  describe('resolve', () => {
    it('should return IBonReceptionItem returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBonReceptionItem = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBonReceptionItem).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBonReceptionItem = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultBonReceptionItem).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IBonReceptionItem>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBonReceptionItem = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBonReceptionItem).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
